refactor: Avoid v0 import when proto v1 is active.
